@jinyus 2023-10-17
Main:

```
 60.57 ms 
906.97 ms 
8.02 s
```

This PR:

```
Crystal:

        Benchmark 1: ./crystal
        Processing time (w/o IO): 60.233623ms
        Processing time (w/o IO): 60.121321ms
        Processing time (w/o IO): 60.199423ms
        Processing time (w/o IO): 60.66213ms
        Processing time (w/o IO): 60.412426ms
        Processing time (w/o IO): 60.191022ms
        Processing time (w/o IO): 60.043319ms
        Processing time (w/o IO): 60.763731ms
        Processing time (w/o IO): 60.145922ms
        Processing time (w/o IO): 60.336124ms
        Processing time (w/o IO): 60.06852ms
        Processing time (w/o IO): 60.572829ms
        Processing time (w/o IO): 59.977818ms
          Time (mean ± σ):     109.3 ms ±   1.3 ms    [User: 97.9 ms, System: 11.4 ms]
          Range (min … max):   107.3 ms … 112.2 ms    10 runs
         
Crystal:

        Benchmark 1: ./crystal
        Processing time (w/o IO): 914.926748ms
        Processing time (w/o IO): 900.691246ms
        Processing time (w/o IO): 906.095438ms
          Time (mean ± σ):      1.108 s ±  0.004 s    [User: 1.086 s, System: 0.022 s]
          Range (min … max):    1.105 s …  1.111 s    2 runs
         
Crystal:

        Benchmark 1: ./crystal
        Processing time (w/o IO): 8002.581562ms
        Processing time (w/o IO): 7969.467772ms
        Processing time (w/o IO): 8017.245475ms
          Time (mean ± σ):      8.721 s ±  0.030 s    [User: 8.611 s, System: 0.102 s]
          Range (min … max):    8.700 s …  8.742 s    2 runs
         

```